Extracted view from activity to fragment
