+++ app/src/main/res/layout/fragment_main.xml
@@ -1,0 +1,112 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout
+  xmlns:android="http://schemas.android.com/apk/res/android"
+  xmlns:app="http://schemas.android.com/apk/res-auto"
+  xmlns:tools="http://schemas.android.com/tools"
+  android:id="@+id/activity_main"
+  android:layout_width="match_parent"
+  android:layout_height="match_parent"
+  tools:context=".mainactivity.MainActivity"
+  >
+
+  <android.support.v4.view.ViewPager
+    android:id="@+id/pager_now_playing"
+    android:layout_width="0dp"
+    android:layout_height="@dimen/height_pager"
+    android:layoutAnimation="@anim/item_now_playing_animation"
+    android:paddingLeft="@dimen/margin_material"
+    android:paddingRight="@dimen/margin_material"
+    app:layout_constraintLeft_toLeftOf="parent"
+    app:layout_constraintRight_toRightOf="parent"
+    app:layout_constraintTop_toTopOf="parent"
+    tools:background="@tools:sample/backgrounds/scenic"
+    />
+
+  <TextView
+    android:id="@+id/movie_title"
+    android:layout_width="0dp"
+    android:layout_height="48dp"
+    android:ellipsize="end"
+    android:fontFamily="sans-serif-condensed"
+    android:gravity="center_vertical"
+    android:maxLines="1"
+    android:textColor="@android:color/white"
+    android:textSize="@dimen/text_size_1"
+    android:visibility="visible"
+    tools:ignore="RtlSymmetry"
+    tools:text="@string/title_now_playing"
+    android:paddingLeft="@dimen/margin_material"
+    android:paddingRight="@dimen/margin_material"
+    app:layout_constraintBottom_toBottomOf="@id/pager_now_playing"
+    app:layout_constraintLeft_toLeftOf="@id/pager_now_playing"
+    app:layout_constraintRight_toRightOf="@id/pager_now_playing"
+    />
+
+  <ImageView
+    android:id="@+id/shader"
+    android:layout_width="0dp"
+    android:layout_height="@dimen/height_shader"
+    android:alpha="0.4"
+    android:background="@drawable/shader"
+    app:layout_constraintLeft_toLeftOf="parent"
+    app:layout_constraintRight_toRightOf="parent"
+    app:layout_constraintTop_toTopOf="parent"
+    />
+
+  <TextView
+    android:id="@+id/title_coming_soon"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginTop="@dimen/margin_material"
+    android:paddingStart="@dimen/margin_material"
+    android:text="@string/title_coming_soon"
+    android:textColor="@android:color/white"
+    android:textSize="@dimen/text_size_2"
+    app:layout_constraintLeft_toLeftOf="parent"
+    app:layout_constraintTop_toBottomOf="@id/pager_now_playing"
+    tools:ignore="RtlSymmetry"
+    />
+
+  <android.support.v7.widget.RecyclerView
+    android:id="@+id/list_coming_soon"
+    android:layout_width="0dp"
+    android:layout_height="@dimen/height_coming_soon"
+    android:layout_marginTop="@dimen/margin_material_half"
+    android:clipToPadding="false"
+    android:layoutAnimation="@anim/item_coming_soon_animation"
+    android:paddingLeft="@dimen/margin_material"
+    android:paddingRight="@dimen/margin_material"
+    app:layout_constraintBottom_toBottomOf="parent"
+    app:layout_constraintLeft_toLeftOf="parent"
+    app:layout_constraintRight_toRightOf="parent"
+    app:layout_constraintTop_toBottomOf="@id/title_coming_soon"
+    tools:listitem="@layout/item_coming_soon"
+    tools:background="@android:color/darker_gray"
+    />
+
+  <android.support.v7.widget.Toolbar
+    android:id="@+id/toolbar"
+    android:layout_width="0dp"
+    android:layout_height="?attr/actionBarSize"
+    android:background="@android:color/transparent"
+    android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
+    app:layout_constraintLeft_toLeftOf="parent"
+    app:layout_constraintRight_toRightOf="parent"
+    app:layout_constraintTop_toTopOf="parent"
+    app:titleTextColor="@android:color/white"
+    tools:title="title"
+    />
+
+  <ProgressBar
+    android:id="@+id/progressbar"
+    android:layout_width="0dp"
+    android:layout_height="wrap_content"
+    android:indeterminate="true"
+    android:visibility="gone"
+    app:layout_constraintLeft_toLeftOf="parent"
+    app:layout_constraintRight_toRightOf="parent"
+    app:layout_constraintTop_toBottomOf="@id/toolbar"
+    tools:visibility="visible"
+    />
+
+</android.support.constraint.ConstraintLayout>
